Fixed silly oversight. Can now allocate data types that are not the same number of bytes as an int.

+++ arralloc.test.c
@@ -7,10 +7,10 @@
 int main(void){
 
     /* our array pointers */
-    int *array_1d;
-    int **array_2d;
-    int ***array_3d;
-    int ****array_4d;
+    long long *array_1d;
+    long long **array_2d;
+    long long ***array_3d;
+    long long ****array_4d;
 
     /* some iterators */
     int i, j, k, l;
@@ -26,7 +26,7 @@ int main(void){
     fprintf(stderr, "Testing 1d version\n");
 
     /* Allocate the array */
-    array_1d = (int*) arralloc(sizeof(int), 1, 500);
+    array_1d = (long long*) arralloc(sizeof(long long), 1, 500);
 
     /* Set data on the array */
     fprintf(stderr, "Testing setting\n");
@@ -41,7 +41,7 @@ int main(void){
             fprintf(
                 stderr,
                 "Incorrect value: array_1d[%d] is %d. Should be %d\n",
-                i, array_1d[i], i
+                i, (int)array_1d[i], i
             );
 
             exit_value = EXIT_FAILURE;
@@ -62,7 +62,7 @@ int main(void){
     fprintf(stderr, "Testing 2d version\n");
 
     /* Allocate the array */
-    array_2d = (int**) arralloc(sizeof(int), 2, 5,100);
+    array_2d = (long long**) arralloc(sizeof(long long), 2, 5,100);
 
     /* Set data on the array */
     fprintf(stderr, "Testing setting\n");
@@ -81,7 +81,7 @@ int main(void){
             fprintf(
                 stderr,
                 "Incorrect value: array_2d[%d][%d] is %d. Should be %d\n",
-                i, j, array_2d[i][j], i*100+j
+                i, j, (int)array_2d[i][j], i*100+j
             );
 
             exit_value = EXIT_FAILURE;
@@ -92,7 +92,7 @@ int main(void){
             fprintf(
                 stderr,
                 "Incorrect value: (&array_2d[0][0])[%d] is %d. Should be %d\n",
-                i*100+j, (&array_2d[0][0])[i*100+j], i*100+j
+                i*100+j, (int)(&array_2d[0][0])[i*100+j], i*100+j
             );
 
             exit_value = EXIT_FAILURE;
@@ -113,7 +113,7 @@ int main(void){
     fprintf(stderr, "Testing 3d version\n");
 
     /* Allocate the array */
-    array_3d = (int***) arralloc(sizeof(int), 3, 5,100,88);
+    array_3d = (long long***) arralloc(sizeof(long long), 3, 5,100,88);
 
     /* Set data on the array */
     fprintf(stderr, "Testing setting\n");
@@ -134,7 +134,7 @@ int main(void){
             fprintf(
                 stderr,
                 "Incorrect value: array_3d[%d][%d][%d] is %d. Should be %d\n",
-                i, j, k, array_3d[i][j][k], i*100*88 +j*88 + k
+                i, j, k, (int)array_3d[i][j][k], i*100*88 +j*88 + k
             );
 
             exit_value = EXIT_FAILURE;
@@ -148,7 +148,7 @@ int main(void){
                 "Should be %d\n",
 
                 i*100*88 +j*88 + k,
-                (&array_3d[0][0][0])[i*100*88 +j*88 + k],
+                (int)(&array_3d[0][0][0])[i*100*88 +j*88 + k],
                 i*100*88 +j*88 + k
             );
 
@@ -170,7 +170,7 @@ int main(void){
     fprintf(stderr, "Testing 4d version\n");
 
     /* Allocate the array */
-    array_4d = (int****) arralloc(sizeof(int), 4, 5,100,88,7);
+    array_4d = (long long****) arralloc(sizeof(long long), 4, 5,100,88,7);
 
     /* Set data on the array */
     fprintf(stderr, "Testing setting\n");
@@ -195,7 +195,7 @@ int main(void){
                 "Incorrect value: array_4d[%d][%d][%d][%d] is %d. "
                 "Should be %d\n",
 
-                i, j, k, l, array_4d[i][j][k][l],
+                i, j, k, l, (int)array_4d[i][j][k][l],
                 i*100*88*7 +j*88*7 + k*7 + l
             );
 
@@ -213,7 +213,7 @@ int main(void){
                 "Should be %d\n",
 
                 i*100*88*7 +j*88*7 + k*7 + l,
-                (&array_4d[0][0][0][0])[i*100*88*7 +j*88*7 + k*7 + l],
+                (int)(&array_4d[0][0][0][0])[i*100*88*7 +j*88*7 + k*7 + l],
                 i*100*88*7 +j*88*7 + k*7 + l
             );
